led_control/requirements.txt
streamlit
rpi_ws281x
pillow
